Refactor Main into Modules (#9)
* Make a Blender env

* Refactor file paths in main.py

* Add Activation Progress Updates

* Update Blender setup in activation script and update .gitignore for zip files

* Refactor README and add environment setup script for Mac/Linux

* Fix Blender download command and update Python executable path in activation script for Linux

* Update .gitignore to correct Blender file patterns and clarify large media files section

* Split main.py into modules

* Refactor render output directory path to use PROJECT_ROOT

* Add gitignore

* Refactor Activate into Setup and Activate

* Enhance activation script to auto-sync dependencies and improve error handling

* Refactor activation script to improve dependency syncing and error handling

* Refactor activation script and add setup script for environment bootstrap for Linux/MacOS

* Update setup and activate after Linux testing

* Refactor README to clarify setup and activation process for Windows and MacOS/Linux

* Fix sourcing setup.sh and activate.sh in parent project folder and env folder

* Enhance setup script with color-coded error messages and improved cleanup handling

* Track Blender files with Git LFS

* Add Earth render blend file

---------
