+++ .gitattributes
@@ -1,0 +1,1 @@
+*.blend filter=lfs diff=lfs merge=lfs -text
