+++ .gitignore
@@ -17,6 +17,7 @@ __pycache__/
 # Blender
 *.blend
 *.blend1
+!assets/Earth.blend
 renders/
 
 # Environment
