+++ assets/Earth.blend
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:a2d34fef38a05001b80f6a84ef887a1bc6f169e9a191ce66b5f99ba6a447e69d
+size 107565719
